Move old reports to one component up:
