+++ src/app/dynamic/enterprise/metering/schedule-config/template.html
@@ -142,8 +142,3 @@ END OF TERMS AND CONDITIONS
 
   </mat-card-content>
 </mat-card>
-
-<ng-container *ngIf="loadOldReportsCard">
-  <km-old-metering-reports-card [reports]="oldReports"
-                                [isLoadingReports]="false"></km-old-metering-reports-card>
-</ng-container>
